v4.1: 修复bug，提交高级语言程序设计大作业实验报告
- 修复灵敏度类型左键切换时取模结果为负数的问题
- 修复三小球/跟枪模式重新进入时状态不重置导致无法开始新游戏的问题
- 修复KPM计算整数除法截断导致数值虚高的问题
- 添加高级语言程序设计大作业实验报告

+++ threeballs.cpp
@@ -137,7 +137,7 @@ void drawthreeballs()
         now_time = GetTickCount() - threeballs_start_time;
         if (now_time > 1000)
         {
-            KPM = score * 60 / (now_time / 1000);
+            KPM = (int)(score * 60.0f / (now_time / 1000.0f));
         }
         else
         {
